Generalise derived field output
Provide an enum of possible derived field variables for equations sets
and routines for setting up the derived field variables for an equations
set.  These allow calculation of a derived variable on demand or for the
derived variable to be used by an equations set, eg. for finite
elasticity to use the strain field with a CellML based constitutive
relation.  Currently only output of a strain field on demand is
implemented.

Conflicts:

	src/equations_routines.f90
	src/equations_set_constants.f90
	src/equations_set_routines.f90
	src/finite_elasticity_routines.f90
	src/opencmiss.f90